# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quant Agent Data Gateway",
    "description": "Quant Agent 执行引擎 HTTP/WS/SSE 契约（BE-19）。\n\n## 统一响应\n\n```json\n{\"code\": 0, \"msg\": \"ok\", \"data\": {}, \"ts\": 1719475200000}\n```\n\n- `code=0` 成功；非零见错误码（`docs/10` §1.4）\n- 响应头 `X-Trace-Id` 用于全链路追踪（BE-10）\n\n## 机器可读契约\n\n导出产物：`docs/openapi.json`（`python scripts/export_openapi.py`）。\n人工互校文档：`docs/10. API接口规范.md`。",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/datalake/snapshots": {
      "get": {
        "tags": [
          "Data Lake Snapshots"
        ],
        "summary": "List Snapshots",
        "operationId": "list_snapshots_api_v1_datalake_snapshots_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "published",
              "title": "Status"
            }
          },
          {
            "name": "storage_tier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Storage Tier"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        },
        "description": "List Snapshots\n\n响应遵循统一信封 `{code, msg, data, ts}`；失败时参见错误码表（docs/10）。"
      }
    },
...
    "/api/v1/datalake/snapshots/{snapshot_id}": {
      "get": {
        "tags": [
          "Data Lake Snapshots"
        ],
        "summary": "Get Snapshot",
        "operationId": "get_snapshot_api_v1_datalake_snapshots__snapshot_id__get",
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        },
        "description": "Get Snapshot\n\n响应遵循统一信封 `{code, msg, data, ts}`；失败时参见错误码表（docs/10）。"
      }
    },
    "/api/v1/datalake/snapshots/rebuild": {
      "post": {
        "tags": [
          "Data Lake Snapshots"
        ],
        "summary": "Rebuild Snapshot",
        "operationId": "rebuild_snapshot_api_v1_datalake_snapshots_rebuild_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RebuildRequest"
              },
              "example": {
                "_note": "见 Schema 字段说明；统一响应见 {code,msg,data,ts}"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        },
        "description": "Rebuild Snapshot\n\n响应遵循统一信封 `{code, msg, data, ts}`；失败时参见错误码表（docs/10）。"
      }
    },
...
    "/api/v1/datasource/{name}/rate-limit-analysis": {
      "get": {
        "tags": [
          "DataSource Rate Limit"
        ],
        "summary": "Get Rate Limit Analysis",
        "description": "查询指定数据源的限流频率分析结果。\n\n返回:\n- estimated_limit_rpm:      推测的限流阈值 RPM\n- recommended_interval_seconds: 推荐安全请求间隔\n- peak_hours:               限流高峰时段\n- avg_recovery_seconds:     平均恢复时间\n- confidence:               推测可信度 (0~1)\n- history:                  每小时统计明细",
        "operationId": "get_rate_limit_analysis_api_v1_datasource__name__rate_limit_analysis_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "分析窗口，如 24h, 7d。默认 24h",
              "examples": [
                "24h",
                "7d"
              ],
              "title": "Window"
            },
            "description": "分析窗口，如 24h, 7d。默认 24h"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasource/{name}/rate-limit-status": {
      "get": {
        "tags": [
          "DataSource Rate Limit"
        ],
        "summary": "Get Rate Limit Status",
        "description": "查询指定数据源的实时退避状态。\n\n返回:\n- is_throttled:           是否处于退避期\n- throttle_until:         退避截止时间戳\n- consecutive_rate_limits: 连续限流次数\n- estimated_rpm:          当前有效 RPM\n- backoff_strategy:       退避策略",
        "operationId": "get_rate_limit_status_api_v1_datasource__name__rate_limit_status_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RebuildRequest": {
        "properties": {
          "as_of_date": {
            "type": "string",
            "title": "As Of Date",
            "description": "YYYY-MM-DD"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "as_of_date"
        ],
        "title": "RebuildRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}